+++ R/get_fy_start.R
@@ -20,9 +20,8 @@
 #' to financial year start. Default is FALSE.
 #' @param q1_is_jan_to_mar boolean. TRUE if Jan to Mar is Q1. FALSE if Apr to
 #' Jun is Q1. Default is FALSE.
-#' @param quarter_pattern character string. Regular expression used to
-#' identify the column that contains only the quarter (e.g. 'Q1' or '1'). Not
-#' required if quarter_from_pattern is supplied.
+#' @param quarter_col_name character string. Exact name of the column that
+#' contains only the quarter number ('1', '2', '3', '4', or NA).
 #' @param month_pattern character string. Regular expression used to
 #' identify the column that contains only the month.
 #'
@@ -42,18 +41,7 @@
 #'     fy_end_pattern = "fy"
 #'     )
 #'
-#' # 2. Only calendar year and quarter are available in a single column.
-#' dat <- data.frame(
-#'     year = c("2021 (Q4 jan-mar of 20-21)", "2021 (Q1 apr-jun of 21-22"),
-#'     value = c(1, 2)
-#'     )
-#' get_fy_start(
-#'     dat, year_from_pattern = "year", get_quarter = TRUE,
-#'     quarter_from_pattern = "year", quarter_col_name = "quarter",
-#'     calendar_to_fy_start = TRUE,  q1_is_jan_to_mar = FALSE,
-#'     )
-#'
-#' # 3. Only calendar year and month are available in separate columns.
+#' # 2. Only calendar year and month are available in separate columns.
 #' dat <- data.frame(
 #'     year = c("2021", "2021"),
 #'     month = c("jan-mar", "apr-jun"),
@@ -69,7 +57,7 @@ get_fy_start <- function(
     dat, year_from_pattern = NA, fy_from_fy_end = FALSE,
     fy_start_from_fy_end = FALSE, fy_end_pattern = NA,
     calendar_to_fy_start = FALSE, q1_is_jan_to_mar = FALSE,
-    quarter_pattern = NA, month_pattern = NA
+    quarter_col_name = NA, month_pattern = NA
 ) {
 
   year_col_created <- dat %>%
@@ -81,7 +69,7 @@ get_fy_start <- function(
   output <- year_col_created %>%
     get_fy_start_from_calendar_year(
       calendar_to_fy_start, q1_is_jan_to_mar, "^year$", month_pattern,
-      quarter_pattern
+      quarter_col_name
     )
 
   return(output)
@@ -93,7 +81,7 @@ get_fy_start <- function(
 #'
 #' @description Find the column containing year using a regular expression,
 #' get the year from that column and put it in a new column called 'year'.
-#' 
+#'
 #' @details
 #' 'year' can contain either financial year, or calendar year,
 #' but not both. By default, if any of the years found in the specified column
@@ -420,8 +408,8 @@ get_fy_from_fy_end <- function(dat, end_year) {
 #' @title Get the year column name to be used by get_quarter
 #'
 #' @description Get the name of the column that matches the year pattern and
-#' contains a valid years. 
-#' 
+#' contains a valid years.
+#'
 #' @details
 #' If more than one year type is found, this function preferentially
 #' returns calendar year, then the column that contains at least some calendar
@@ -491,9 +479,9 @@ get_year_col_to_drop_from_quarter <- function(dat, pattern) {
 #'
 #' @details
 #' Where the output of tidysheet is used for calculations on a financial year
-#' basis, but the source data was published on a calendar year basis (where Q1 
-#' is the first quarter of the calendar year), financial year may not be given. 
-#' This function calculates financial year start from the calendar year and 
+#' basis, but the source data was published on a calendar year basis (where Q1
+#' is the first quarter of the calendar year), financial year may not be given.
+#' This function calculates financial year start from the calendar year and
 #' either the quarter or the month.
 #'
 #' See make_calendar_q1_month_and_quarter_patterns() for the accepted
@@ -517,9 +505,9 @@ get_year_col_to_drop_from_quarter <- function(dat, pattern) {
 #' will be seen as '2024'). There must be no numbers present other than the
 #' year. Column can be character or integer.
 #' @param month_col_pattern character string. Same as year_from_pattern but for
-#' month. Only required if quarter_col_pattern is not given.
-#' @param quarter_col_pattern character string. Same as year_from_pattern but for
-#' quarter. Only required if month_col_pattern is not given.
+#' month. Only required if quarter_col_name is not given.
+#' @param quarter_col_name character string. Name of the coloumn containing the
+#' quarter number. Only required if month_col_pattern is not given.
 #' @param calendar_year_to_fy_start boolean. Default is FALSE. TRUE if the
 #' action is to be performed, NA or FALSE to return the original data.
 #'
@@ -539,7 +527,7 @@ get_year_col_to_drop_from_quarter <- function(dat, pattern) {
 #'     dat, calendar_year_to_fy_start = TRUE,
 #'     q1_is_jan_to_mar = FALSE,
 #'     year_from_pattern = "(?i)year",
-#'     quarter_col_pattern = "(?i)quarter",
+#'     quarter_col_name = "Quarter",
 #'     month_col_pattern = NA
 #'     )
 #'
@@ -554,7 +542,7 @@ get_year_col_to_drop_from_quarter <- function(dat, pattern) {
 #'     dat, calendar_year_to_fy_start = TRUE,
 #'     q1_is_jan_to_mar = FALSE,
 #'     year_from_pattern = "(?i)year",
-#'     quarter_col_pattern = "(?i)quarter"
+#'     quarter_col_name = "Quarter"
 #'     )
 #'
 #' # month is given for some rows, quarter is given for others, and some have both:
@@ -568,7 +556,7 @@ get_year_col_to_drop_from_quarter <- function(dat, pattern) {
 #'     dat, calendar_year_to_fy_start = TRUE,
 #'     q1_is_jan_to_mar = FALSE,
 #'     year_from_pattern = "(?i)year",
-#'     quarter_col_pattern = "(?i)quarter",
+#'     quarter_col_name = "Quarter",
 #'     month_col_pattern = "(?i)month"
 #'     )
 #' }
@@ -576,7 +564,7 @@ get_year_col_to_drop_from_quarter <- function(dat, pattern) {
 get_fy_start_from_calendar_year <- function(
     dat, calendar_year_to_fy_start = FALSE, q1_is_jan_to_mar = TRUE,
     year_from_pattern = "(?i)year", month_col_pattern = NA,
-    quarter_col_pattern = NA
+    quarter_col_name = NA
 ) {
 
   if (is.na(calendar_year_to_fy_start) | calendar_year_to_fy_start == FALSE) {
@@ -585,11 +573,11 @@ get_fy_start_from_calendar_year <- function(
 
   message("Getting financial year start from column containing calendar year.")
 
-  if (all(is.na(month_col_pattern), is.na(quarter_col_pattern),
+  if (all(is.na(month_col_pattern), is.na(quarter_col_name),
           calendar_year_to_fy_start == TRUE)) {
     stop("calendar_year_to_fy_start is TRUE, but neither ",
-         "month_col_pattern, nor quarter_col_pattern has been supplied. ",
-         "Developer: please supply one of these in the data dictionary.")
+         "month_col_pattern, nor quarter_col_name has been supplied. ",
+         "Please supply one of these in the settings.")
   }
 
   if ("fy_start" %in% names(dat)) {
@@ -625,19 +613,32 @@ get_fy_start_from_calendar_year <- function(
     )
   }
 
+  # We used to use a user supplied pattern for finding the name of the quarter
+  # column, but that is now replaced with a known name in add_quarter_column so
+  # we can use the known name. So we change as little code as possible, we can
+  # just make the known name a pattern by tying it to the start (^) and end($):
+  if (!is.na(quarter_col_name)) {
+     quarter_col_pattern <- paste0("^", quarter_col_name, "$")
+  } else {
+    quarter_col_pattern <- NA
+  }
+
+
   period_col_names <- get_colnames_from_pattern(
     dat, c("month", "quarter"), c(month_col_pattern, quarter_col_pattern)
   )
+
   if (!is.na(month_col_pattern) & is.na(period_col_names["month"])) {
     stop(
       "Please fix the settings to address the issue with month_col_pattern. ",
       "month_col_pattern and/or calendar_year_to_fy_start may need to change."
     )
   }
-  if (!is.na(quarter_col_pattern) & is.na(period_col_names["quarter"])) {
+  if (!is.na(quarter_col_name) & is.na(period_col_names["quarter"])) {
     stop(
-      "Please fix the settings to address the issue with quarter_col_pattern. ",
-      "quarter_col_pattern and/or calendar_year_to_fy_start may need to change."
+      "Please fix the settings to address the issue with quarter_col_name. ",
+      "quarter_from_col_pattern and/or calendar_year_to_fy_start may need to ",
+      "change."
     )
   }