+++ R/process_year_column.R
@@ -60,14 +60,18 @@
 #'
 #' @examples
 #' \dontrun{
-#' dat <- data.frame(year_col = c("2021", "2022_23 to 2025_26", "2021, 2022"))
+#' dat <- data.frame(
+#'    year_col = c("2021", "2022_23 to 2025_26", "2021, 2022"),
+#'    is_blank = FALSE
+#'    )
 #' process_year_column(dat, "year_col", NA, NA, NA, NA)
 #'
 #' dat <- data.frame(
 #'     financial_year = c("2021-22 to 2022-23", "2022_2023", "words"),
-#'     calendar_year = c(2021, 2022, NA)
+#'     calendar_year = c(2021, 2022, NA),
+#'     is_blank = FALSE
 #'     )
-#' process_year_column(dat, NA, NA, NA, NA, TRUE)
+#' process_year_column(dat, NA, FALSE, NA, NA, TRUE)
 #' process_year_column(dat, NA, NA, NA, NA, FALSE)
 #'
 #' # case where there is no year column, but there is a column that looks like
@@ -75,7 +79,8 @@
 #' dat <- data.frame(
 #'     description = c("A", "B", "C"),
 #'     implementation_year = c(1980, 1992, 2003),
-#'     value = 1:3
+#'     value = 1:3,
+#'     is_blank = FALSE
 #'     )
 #' single_year = list("year" = "2021_22", "warn" = TRUE)
 #' process_year_column(